backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
openai==1.10.0
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-core==1.29.5
azure-keyvault-secrets==4.8.0
python-multipart==0.0.6
